+++ R/validate_date.R
@@ -63,9 +63,10 @@ validate_date <- function(
       validate_error_type(
         input = input_name,
         message = "must not be NULL.",
-        type = "e"
+        type = "error"
       )
     }
+    return(NULL)
   }
 
   # Check if the input is a valid date