+++ R/validate_numeric.R
@@ -73,9 +73,10 @@ validate_numeric <- function(
       validate_error_type(
         input = input_name,
         message = "must not be NULL.",
-        type = "e"
+        type = "error"
       )
     }
+    return(NULL)
   }
 
   # Check if the input is numeric